Consolidate code files into internal package
